gather programs' method and cbs lists into graft.py
