+++ qcdb/driver/load_proc_table.py
@@ -4,7 +4,7 @@ from qcengine.programs import register_program
 
 from ..programs.cfour import QcdbCFOURHarness, cfour_gradient_list, cfour_hessian_list, cfour_list, run_cfour
 from ..programs.dftd3 import alt_run_dftd3, dftd3_list
-from ..programs.gamess import QcdbGAMESSHarness, gamess_gradient_list, gamess_list, run_gamess
+from ..programs.gamess import QcdbGAMESSHarness, gamess_gradient_list, gamess_hessian_list, gamess_list, run_gamess
 from ..programs.nwchem import QcdbNWChemHarness, nwchem_gradient_list, nwchem_hessian_list, nwchem_list, run_nwchem
 from ..programs.psi4 import QcdbPsi4Harness, run_psi4
 from .proc_table import procedures
@@ -85,7 +85,10 @@ if which('rungms'):
     for mtd in gamess_list():
         procedures['energy']['gamess'][mtd.lower()] = run_gamess
 
-    # integrate GAMESS with driver routines
     procedures['gradient']['gamess'] = {}
     for mtd in gamess_gradient_list():
         procedures['gradient']['gamess'][mtd.lower()] = run_gamess
+
+    procedures['hessian']['gamess'] = {}
+    for mtd in gamess_hessian_list():
+        procedures['hessian']['gamess'][mtd.lower()] = run_gamess
