+++ qcdb/programs/cfour/__init__.py
@@ -1,2 +1,2 @@
-from .harvester import cfour_gradient_list, cfour_hessian_list, cfour_list
+from .graft import cfour_gradient_list, cfour_hessian_list, cfour_list
 from .runner import QcdbCFOURHarness, run_cfour
