+++ qcdb/programs/gamess/__init__.py
@@ -1,2 +1,2 @@
-from .plant import gamess_gradient_list, gamess_list
+from .graft import gamess_gradient_list, gamess_hessian_list, gamess_list
 from .runner import QcdbGAMESSHarness, run_gamess
