+++ qcdb/programs/gamess/runner.py
@@ -15,104 +15,14 @@ from qcengine.programs.util import PreservingDict
 
 from ... import qcvars
 from ...basisset import BasisSet
-from ...util import provenance_stamp
+from ...util import print_jobrec, provenance_stamp
 from .harvester import muster_inherited_options, muster_modelchem
 from .molbasopt import muster_and_format_molecule_and_basis_for_gamess
 
 pp = pprint.PrettyPrinter(width=120)
 
 
-
-
-
-#def run_gamess_old(name, molecule, options, **kwargs):
-#
-#    prov = {}
-#    prov['creator'] = 'QCDB'
-#    prov['version'] = __version__
-#    prov['routine'] = sys._getframe().f_code.co_name
-#    
-#
-#    qcschema_input = {
-#        'schema_name': 'qcschema_input',
-#        'schema_version': 1,
-#        'driver': inspect.stack()[1][3],
-#        'model': {
-#            'method': name,
-#            'basis': '(auto)',
-#        },
-#        'molecule': molecule.to_schema(dtype=2),
-#        'extras': {},
-#    }
-#
-#    jobrec = {}
-#    jobrec['provenance'] = [prov]
-#    jobrec['error'] = ''
-#    jobrec['success'] = None
-#    jobrec['molecule'] = molecule.to_dict(np_out=False)
-#    jobrec['method'] = name
-#    jobrec['dertype'] = ['energy', 'gradient', 'hessian'].index(inspect.stack()[1][3])
-#
-#
-#    jobrec['options'] = copy.deepcopy(options)
-#    jobrec['qcschema_input'] = qcschema_input
-#
-#    gamessrec = {}
-#    muster_inherited_options(jobrec['options'])
-#
-#    qmol = Molecule(jobrec['molecule'])
-#    _qcdb_basis = jobrec['options'].scroll['QCDB']['BASIS'].value
-#    _gamess_basis = jobrec['options'].scroll['GAMESS']['BASIS__GBASIS'].value
-#    if _qcdb_basis == '':
-#        raise ValueError('gamess bas not impl')
-#    qbs = BasisSet.pyconstruct(jobrec['molecule'], 'BASIS', _qcdb_basis)
-#
-#    molbascmd = muster_and_format_molecule_and_basis_for_gamess(jobrec['molecule'], jobrec['options'], qbs, verbose=1)
-#
-#
-#    # Handle calc type and quantum chemical method
-#    nel = sum([z * int(real) for z, real in zip(jobrec['molecule']['elez'], jobrec['molecule']['real'])]) - jobrec['molecule']['molecular_charge']
-#    nel = int(nel)
-#    nfzc = qmol.nfrozen_core(depth=True)  # this will be default FC  # TODO change these values when user sets custom FC
-#    nfzc = 0
-#
-#    # forcing nfc above. all these need to be reocmputed together for a consistent cidet input group
-#    nels = nel - 2 * nfzc
-#    nact = qbs.nbf() - nfzc
-#    sysinfo = {
-#        'nel': nel,
-#        'ncore': nfzc,
-#        'nact': nact,
-#        'nels': nels,
-#    }
-#
-#    muster_modelchem(jobrec['method'], jobrec['dertype'], jobrec['options'], sysinfo)
-#
-#    # Handle conversion of qcdb keyword structure into gamess format
-#    resolved_options = {k: v.value for k, v in jobrec['options'].scroll['GAMESS'].items() if v.disputed()}
-#    optcmd = format_options_for_gamess(resolved_options)
-#
-#    # Assemble input pieces
-#    gamessrec['gamess.inp'] = optcmd + molbascmd
-#
-#    qcschema_input = jobrec['qcschema_input']
-#    qcschema_input['extras']['gamess.inp'] = gamessrec['gamess.inp']
-#
-#
-#    ret = qcng.compute(qcschema_input, 'gamess').dict()
-#    jobrec.pop('qcschema_input')
-#    progvars = PreservingDict(ret['extras']['qcvars'])
-#    qcvars.build_out(progvars)
-#    calcinfo = qcvars.certify_and_datumize(progvars, plump=True, nat=len(jobrec['molecule']['mass']))
-#    jobrec['raw_output'] = ret['stdout']
-#    jobrec['qcvars'] = calcinfo
-#    jobrec['success'] = True
-#
-#    return jobrec
-
-
 def run_gamess(name, molecule, options, **kwargs):
-#def run_gamess2(name, molecule, options, **kwargs):
 
     resi = ResultInput(
         **{
@@ -133,12 +43,6 @@ def run_gamess(name, molecule, options, **kwargs):
     jobrec['qcvars'] = {key: qcel.Datum(**dval) for key, dval in hold_qcvars.items()}
     return jobrec
 
-def _print_helper(label, dicary, do_print):
-    if do_print:
-        print(label + ' <<<')
-        pp.pprint(dicary)
-        print('>>>')
-
 
 class QcdbGAMESSHarness(GAMESSHarness):
     def compute(self, input_model: 'ResultInput', config: 'JobConfig') -> 'Result':
@@ -146,15 +50,15 @@ class QcdbGAMESSHarness(GAMESSHarness):
 
         verbose = 1
 
-        _print_helper(f'[1] {self.name} RESULTINPUT PRE-PLANT', input_model.dict(), verbose >= 3)
+        print_jobrec(f'[1] {self.name} RESULTINPUT PRE-PLANT', input_model.dict(), verbose >= 3)
 
         job_inputs = self.qcdb_build_input(input_model, config)
 
-        _print_helper(f'[2] {self.name}REC PRE-ENGINE', job_inputs, verbose >= 4)
+        print_jobrec(f'[2] {self.name}REC PRE-ENGINE', job_inputs, verbose >= 4)
 
         success, dexe = self.execute(job_inputs)
 
-        _print_helper(f'[3] {self.name}REC POST-ENGINE', dexe, verbose >= 4)
+        print_jobrec(f'[3] {self.name}REC POST-ENGINE', dexe, verbose >= 4)
 
         if 'INPUT HAS AT LEAST ONE SPELLING OR LOGIC MISTAKE' in dexe["stdout"]:
             raise InputError(dexe["stdout"])
@@ -167,11 +71,11 @@ class QcdbGAMESSHarness(GAMESSHarness):
         dexe["outfiles"]["stderr"] = dexe["stderr"]
         output_model = self.parse_output(dexe["outfiles"], input_model)
 
-        _print_helper(f'[4a] {self.name} RESULT POST-HARVEST', output_model.dict(), verbose >= 5)
+        print_jobrec(f'[4a] {self.name} RESULT POST-HARVEST', output_model.dict(), verbose >= 5)
 
         output_model = self.qcdb_post_parse_output(input_model, output_model)
 
-        _print_helper(f'[4] {self.name} RESULT POST-POST-HARVEST', output_model.dict(), verbose >= 2)
+        print_jobrec(f'[4] {self.name} RESULT POST-POST-HARVEST', output_model.dict(), verbose >= 2)
 
         return output_model
 
@@ -234,100 +138,3 @@ class QcdbGAMESSHarness(GAMESSHarness):
         output_model.extras['qcdb:qcvars'] = calcinfo
 
         return output_model
-
-
-
-#def cfour_harvest(jobrec, cfourrec):  # jobrec@i, cfourrec@io -> jobrec@io
-#    """Processes raw results from read-only `cfourrec` into QCAspect fields in returned `jobrec`."""
-#
-#    try:
-#        pass
-#        #jobrec['molecule']['real']
-#        #jobrec['do_gradient']
-#    except KeyError as err:
-#        raise KeyError(
-#            'Required fields missing from ({})'.format(jobrec.keys())) from err
-#
-#    try:
-#        cfourrec['stdout']
-#        #if jobrec['do_gradient'] is True:
-#        #    dftd3rec['dftd3_gradient']
-#    except KeyError as err:
-#        raise KeyError('Required fields missing from ({})'.format(
-#            cfourrec.keys())) from err
-#
-#    # amalgamate output
-#    text = cfourrec['stdout']
-#    text += '\n  <<<  Cfour {} {} Results  >>>\n\n'.format('', '') #name.lower(), calledby.capitalize()))  # banner()
-#
-#    c4files = {}
-#    for fl in ['GRD', 'FCMFINAL', 'DIPOL']:
-#        field = 'output_' + fl.lower()
-#        if field in cfourrec:
-#            text += '  Cfour scratch file {} has been read\n'.format(fl)
-#            text += cfourrec[field]
-#            c4files[fl] = cfourrec[field]
-#
-#
-#    #if molecule.name() == 'blank_molecule_psi4_yo':
-#    #    qcdbmolecule = None
-#    #else:
-#    #    molecule.update_geometry()
-#    #    qcdbmolecule = qcdb.Molecule(molecule.create_psi4_string_from_molecule())
-#    #    qcdbmolecule.update_geometry()
-#    qmol = Molecule(jobrec['molecule'])
-#
-#    # c4mol, if it exists, is dinky, just a clue to geometry of cfour results
-#    psivar, c4hess, c4grad, c4mol, version, errorTMP = harvester.harvest(qmol, cfourrec['stdout'], **c4files)
-#
-#    jobrec['error'] += errorTMP
-#    # Absorb results into psi4 data structures
-#    #for key in psivar.keys():
-#    #    core.set_variable(key.upper(), float(psivar[key]))
-#    #calcinfo = qcvars.certify_qcvars(psivar)
-#    #jobrec['qcvars'] = {info.lbl: info for info in calcinfo}
-#    progvars = PreservingDict(psivar)
-#
-#    #if qcdbmolecule is None and c4mol is not None:
-#    #    molecule = geometry(c4mol.create_psi4_string_from_molecule(), name='blank_molecule_psi4_yo')
-#    #    molecule.update_geometry()
-#    #    # This case arises when no Molecule going into calc (cfour {} block) but want
-#    #    #   to know the orientation at which grad, properties, etc. are returned (c4mol).
-#    #    #   c4mol is dinky, w/o chg, mult, dummies and retains name
-#    #    #   blank_molecule_psi4_yo so as to not interfere with future cfour {} blocks
-#
-#    if c4grad is not None:
-#        progvars['CURRENT GRADIENT'] = c4grad
-#        #mat = core.Matrix.from_list(c4grad)
-#        #core.set_gradient(mat)
-#
-#    if c4hess is not None:
-#        progvars['CURRENT HESSIAN'] = c4hess
-#
-#    # badly placed
-#    # Cfour's SCS-MP2 is non adjustible and only valid for UHF
-#    # ROMP2 doesn't print SS & OS
-#    if "MP2 OPPOSITE-SPIN CORRELATION ENERGY" in progvars and "MP2 SAME-SPIN CORRELATION ENERGY" in progvars:
-#        oss_opt = jobrec['options'].scroll['QCDB']['MP2_OS_SCALE']
-#        sss_opt = jobrec['options'].scroll['QCDB']['MP2_SS_SCALE']
-#        custom_scsmp2_corl = \
-#            Decimal(oss_opt.value) * progvars["MP2 OPPOSITE-SPIN CORRELATION ENERGY"] + \
-#            Decimal(sss_opt.value) * progvars["MP2 SAME-SPIN CORRELATION ENERGY"]
-#        if "MP2 SINGLES ENERGY" in progvars:
-#            custom_scsmp2_corl += progvars["MP2 SINGLES ENERGY"]
-#        progvars["CUSTOM SCS-MP2 CORRELATION ENERGY"] = custom_scsmp2_corl
-#
-#    qcvars.build_out(progvars)
-#    calcinfo = qcvars.certify_and_datumize(progvars)
-#    text += print_variables(calcinfo)
-#
-#    jobrec['raw_output'] = text
-#    jobrec['qcvars'] = calcinfo
-#
-#    prov = {}
-#    prov['creator'] = 'Cfour'
-#    prov['routine'] = sys._getframe().f_code.co_name
-#    prov['version'] = version
-#    jobrec['provenance'].append(prov)
-#
-#    return jobrec
