+++ qcdb/programs/nwchem/__init__.py
@@ -1,2 +1,2 @@
-from .harvester import nwchem_gradient_list, nwchem_hessian_list, nwchem_list
+from .graft import nwchem_gradient_list, nwchem_hessian_list, nwchem_list
 from .runner import QcdbNWChemHarness, run_nwchem
